+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- bundle a web emoji font subset for the profile attribution
+
 ## [1.1.0] - 2026-04-10
 ### Added
 - Changelog
@@ -34,5 +38,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - settings
   - login
 
+[Unreleased]: https://github.com/donjoergo/glasstrail/compare/1.1.0...HEAD
 [1.1.0]: https://github.com/donjoergo/glasstrail/compare/1.0.1...1.1.0
 [1.0.1]: https://github.com/donjoergo/glasstrail/releases/tag/1.0.1